CreateRoom
Added a new Scene. Start in "ConnectToLobby" and then press play. You type in your name and press the button. Then you get to a new Scene and you can type in your room name and it gets created.

Added a scrollable view, but there are no rooms displayed yet.